requirements.txt
google-api-python-client>=2.150.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.1
openai>=1.54.0
pyinstaller>=6.11.0
